+++ app/backend/component/gadget.go
@@ -39,6 +39,20 @@ func validateGadgetType(input GadgetType) duerror.DUError {
 	return nil
 }
 
+func validateSection(section, numSections int) duerror.DUError {
+	if section < 0 || section >= numSections {
+		return duerror.NewInvalidArgumentError("section out of range")
+	}
+	return nil
+}
+
+func validateIndex(index, numItems int) duerror.DUError {
+	if index < 0 || index >= numItems {
+		return duerror.NewInvalidArgumentError("index out of range")
+	}
+	return nil
+}
+
 // Constructor
 func NewGadget(gadgetType GadgetType, point utils.Point, layer int, colorHexStr string, header string) (*Gadget, duerror.DUError) {
 	if err := validateGadgetType(gadgetType); err != nil {
@@ -119,10 +133,10 @@ func (g *Gadget) SetColor(colorHexStr string) duerror.DUError {
 }
 
 func (g *Gadget) SetAttrContent(section int, index int, content string) duerror.DUError {
-	if err := utils.ValidateSection(section, len(g.attributes)); err != nil {
+	if err := validateSection(section, len(g.attributes)); err != nil {
 		return err
 	}
-	if err := utils.ValidateIndex(index, len(g.attributes[section])); err != nil {
+	if err := validateIndex(index, len(g.attributes[section])); err != nil {
 		return err
 	}
 	if err := g.attributes[section][index].SetContent(content); err != nil {
@@ -132,10 +146,10 @@ func (g *Gadget) SetAttrContent(section int, index int, content string) duerror.
 }
 
 func (g *Gadget) SetAttrSize(section int, index int, size int) duerror.DUError {
-	if err := utils.ValidateSection(section, len(g.attributes)); err != nil {
+	if err := validateSection(section, len(g.attributes)); err != nil {
 		return err
 	}
-	if err := utils.ValidateIndex(index, len(g.attributes[section])); err != nil {
+	if err := validateIndex(index, len(g.attributes[section])); err != nil {
 		return err
 	}
 	if err := g.attributes[section][index].SetSize(size); err != nil {
@@ -145,10 +159,10 @@ func (g *Gadget) SetAttrSize(section int, index int, size int) duerror.DUError {
 }
 
 func (g *Gadget) SetAttrStyle(section int, index int, style int) duerror.DUError {
-	if err := utils.ValidateSection(section, len(g.attributes)); err != nil {
+	if err := validateSection(section, len(g.attributes)); err != nil {
 		return err
 	}
-	if err := utils.ValidateIndex(index, len(g.attributes[section])); err != nil {
+	if err := validateIndex(index, len(g.attributes[section])); err != nil {
 		return err
 	}
 	if err := g.attributes[section][index].SetStyle(attribute.Textstyle(style)); err != nil {
@@ -165,7 +179,7 @@ func (g *Gadget) Cover(p utils.Point) (bool, duerror.DUError) {
 }
 
 func (g *Gadget) AddAttribute(section int, content string) duerror.DUError {
-	if err := utils.ValidateSection(section, len(g.attributes)); err != nil {
+	if err := validateSection(section, len(g.attributes)); err != nil {
 		return err
 	}
 	att, err := attribute.NewAttribute(content)
@@ -180,10 +194,10 @@ func (g *Gadget) AddAttribute(section int, content string) duerror.DUError {
 }
 
 func (g *Gadget) RemoveAttribute(section int, index int) duerror.DUError {
-	if err := utils.ValidateSection(section, len(g.attributes)); err != nil {
+	if err := validateSection(section, len(g.attributes)); err != nil {
 		return err
 	}
-	if err := utils.ValidateIndex(index, len(g.attributes[section])); err != nil {
+	if err := validateIndex(index, len(g.attributes[section])); err != nil {
 		return err
 	}
 	g.attributes[section] = append(g.attributes[section][:index], g.attributes[section][index+1:]...)
