+++ app/backend/component/gadget_test.go
@@ -376,3 +376,59 @@ func TestRegisterUpdateParentDraw(t *testing.T) {
 	// nil function
 	assert.Error(t, g.RegisterUpdateParentDraw(nil))
 }
+
+func TestValidateSection(t *testing.T) {
+	tests := []struct {
+		name        string
+		section     int
+		numSections int
+		hasError    bool
+	}{
+		{"ValidSection", 0, 3, false},
+		{"ValidSectionMiddle", 1, 3, false},
+		{"ValidSectionLast", 2, 3, false},
+		{"NegativeSection", -1, 3, true},
+		{"SectionEqualToNumSections", 3, 3, true},
+		{"SectionGreaterThanNumSections", 4, 3, true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := validateSection(tt.section, tt.numSections)
+			if tt.hasError {
+				assert.Error(t, err)
+				assert.IsType(t, duerror.NewInvalidArgumentError(""), err)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestValidateIndex(t *testing.T) {
+	tests := []struct {
+		name     string
+		index    int
+		numItems int
+		hasError bool
+	}{
+		{"ValidIndex", 0, 3, false},
+		{"ValidIndexMiddle", 1, 3, false},
+		{"ValidIndexLast", 2, 3, false},
+		{"NegativeIndex", -1, 3, true},
+		{"IndexEqualToNumItems", 3, 3, true},
+		{"IndexGreaterThanNumItems", 4, 3, true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := validateIndex(tt.index, tt.numItems)
+			if tt.hasError {
+				assert.Error(t, err)
+				assert.IsType(t, duerror.NewInvalidArgumentError(""), err)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
